+++ R/match-moments.R
@@ -52,7 +52,7 @@ min_discrepancy <- function(dist, meanlog, sdlog, nsim) {
 #' print(moments)
 #' ssd_hc(moments)
 #' ssd_plot_cdf(moments)
-ssd_match_moments <- function(dists = ssd_dists("bc"), meanlog = 1, sdlog = 1,
+ssd_match_moments <- function(dists = ssd_dists_bcanz(), meanlog = 1, sdlog = 1,
                               nsim = 1e+05) {
   chk_vector(dists)
   chk_s3_class(dists, "character")