+++ data-raw/data-raw.R
@@ -14,7 +14,24 @@
 
 library(ssdtools)
 library(ssddata)
+library(tibble)
 library(usethis)
+
+dist_data <- tibble::tribble(
+  ~dist, ~bcanz, ~tails, ~npars,
+  "burrIII3",  FALSE,   TRUE,     3L,
+  "gamma",   TRUE,   TRUE,     2L,
+  "gompertz",  FALSE,   TRUE,     2L,
+  "invpareto",  FALSE,  FALSE,     2L,
+  "lgumbel",   TRUE,   TRUE,     2L,
+  "llogis",   TRUE,   TRUE,     2L,
+  "llogis_llogis",  FALSE,   TRUE,     5L,
+  "lnorm",   TRUE,   TRUE,     2L,
+  "lnorm_lnorm",   TRUE,   TRUE,     5L,
+  "weibull",   TRUE,   TRUE,     2L
+)
+
+use_data(dist_data, overwrite = TRUE)
 
 boron_data <- ssddata::ccme_boron
 use_data(boron_data, overwrite = TRUE)